requirements.txt
pytesseract==0.3.10
pdf2image==1.16.3
spacy==3.7.2
python-docx==1.0.1
opencv-python==4.8.1.78
Pillow==10.0.1
pandas==2.1.1 